requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
langchain-google-genai
langchain-community
langchain-core
Pillow
pytesseract